refactor: move 'withLogout' to User context
